+++ Packagers/package-freecad.ps1
@@ -7,6 +7,7 @@ CPE: cpe:2.3:a:freecad:freecad:*:*:*:*:*:*:*:*
 ReleaseNotesUrl: https://github.com/FreeCAD/FreeCAD/releases
 DownloadPageUrl: https://www.freecad.org/downloads.php
 IconSource: Installer
+SupportsInstallModes: CurrentUser, AllUsers
 UpdateCadenceDays: 90
 
 .SYNOPSIS