+++ Packagers/package-gimp.ps1
@@ -7,6 +7,7 @@ CPE: cpe:2.3:a:gimp:gimp:*:*:*:*:*:*:*:*
 ReleaseNotesUrl: https://www.gimp.org/news/
 DownloadPageUrl: https://www.gimp.org/downloads/
 IconSource: Installer
+SupportsInstallModes: CurrentUser, AllUsers
 
 .SYNOPSIS
     Packages GIMP (x64) for MECM.