+++ Packagers/package-joplin.ps1
@@ -7,6 +7,7 @@ CPE: cpe:2.3:a:joplinapp:joplin:*:*:*:*:*:*:*:*
 ReleaseNotesUrl: https://joplinapp.org/help/about/changelog/desktop/
 DownloadPageUrl: https://joplinapp.org/download/
 IconSource: Installer
+SupportsInstallModes: CurrentUser, AllUsers
 UpdateCadenceDays: 30
 
 .SYNOPSIS