+++ Packagers/package-kdiff3.ps1
@@ -7,6 +7,7 @@ CPE: cpe:2.3:a:kde:kdiff3:*:*:*:*:*:*:*:*
 ReleaseNotesUrl: https://invent.kde.org/sdk/kdiff3/-/releases
 DownloadPageUrl: https://download.kde.org/stable/kdiff3/
 IconSource: Installer
+SupportsInstallModes: CurrentUser, AllUsers
 UpdateCadenceDays: 180
 
 .SYNOPSIS