+++ Packagers/package-obsidian.ps1
@@ -7,6 +7,7 @@ CPE: cpe:2.3:a:obsidian:obsidian:*:*:*:*:*:*:*:*
 ReleaseNotesUrl: https://obsidian.md/changelog/
 DownloadPageUrl: https://github.com/obsidianmd/obsidian-releases/releases
 IconSource: Installer
+SupportsInstallModes: CurrentUser, AllUsers
 UpdateCadenceDays: 30
 
 .SYNOPSIS