+++ Packagers/package-rstudio.ps1
@@ -7,6 +7,7 @@ CPE: cpe:2.3:a:posit:rstudio:*:*:*:*:*:*:*:*
 ReleaseNotesUrl: https://docs.posit.co/ide/news/
 DownloadPageUrl: https://posit.co/download/rstudio-desktop/
 IconSource: Installer
+SupportsInstallModes: CurrentUser, AllUsers
 
 .SYNOPSIS
     Packages RStudio Desktop (x64) for MECM.