+++ Packagers/package-thonny.ps1
@@ -7,6 +7,7 @@ CPE: cpe:2.3:a:thonny:thonny:*:*:*:*:*:*:*:*
 ReleaseNotesUrl: https://github.com/thonny/thonny/blob/master/CHANGELOG.rst
 DownloadPageUrl: https://github.com/thonny/thonny/releases/latest
 IconSource: External
+SupportsInstallModes: CurrentUser, AllUsers
 UpdateCadenceDays: 90
 
 .SYNOPSIS